+++ .rubocop.yml
@@ -2,17 +2,17 @@ AllCops:
     NewCops: enable
     TargetRubyVersion: 2.0
 Metrics/AbcSize:
-    Max: 35
+    Max: 24
 Metrics/BlockLength:
     Max: 51
 Metrics/CyclomaticComplexity:
-    Max: 8
+    Max: 6
 Metrics/MethodLength:
-    Max: 36
+    Max: 28
 Naming/MethodParameterName:
     Enabled: false
 Metrics/PerceivedComplexity:
-    Max: 7
+    Max: 6
 Naming/VariableNumber:
     Enabled: false
 Style/FrozenStringLiteralComment:
