+++ lib/decrypt.rb
@@ -19,6 +19,7 @@ require 'io/console'
 require 'openssl'
 require 'optparse'
 
+require_relative 'crypto'
 require_relative 'pretty'
 
 ITERATIONS = 10_000
@@ -80,86 +81,6 @@ def split_cipher_text(cipher_text_with_auth_tag)
 end
 
 #
-# Derive a key from the given password and salt using PBKDF2-HMAC.
-#
-# @param [String] password Backup file password as plaintext
-# @param [String] salt HMAC salt as bytes
-#
-# @return [String] Derived key
-#
-def derive_key(password, salt)
-  OpenSSL::PKCS5.pbkdf2_hmac(password, salt, ITERATIONS, KEY_LENGTH / 8,
-                             HASH)
-end
-
-#
-# Perform AES-GCM encryption or decryption.
-#
-# @param [String] text Text to be encrypted or decrypted
-# @param [String] master_key AES-GCM master key
-# @param [String] iv AES-GCM initialization vector
-# @param [Boolean] encrypt Specify whether encryption or decryption should be performed
-# @param [String, nil] auth_tag AES-GCM authentication tag used for decryption. Will not be used if `encrypt` is true.
-#
-# @return [Array<String>] 2-element Array where first element is resulting ciphertext or plaintext, and second element
-#  is the AES-GCM authentication tag
-#
-def aes_gcm(text, master_key, iv, encrypt, auth_tag = nil)
-  cipher = OpenSSL::Cipher.new ENCRYPTION_CIPHER
-  if encrypt
-    cipher.encrypt
-  else
-    cipher.decrypt
-  end
-  cipher.key = master_key
-  cipher.iv = iv
-  cipher.auth_tag = auth_tag unless encrypt
-  cipher.auth_data = ''
-  cipher.padding = 0
-
-  [cipher.update(text) + cipher.final, encrypt ? cipher.auth_tag : auth_tag]
-end
-
-#
-# Decrypt `cipher_text` and return its plaintext and authentication tag.
-#
-# @param [String] cipher_text Encrypted text as bytes to be decrypted
-# @param [String] password Backup file password as plaintext
-# @param [String] salt HMAC salt as bytes
-# @param [String] iv AES-GCM initialization vector as bytes
-# @param [String] auth_tag AES-GCM authentication tag as bytes
-#
-# @return [Array<String>] 2-element Array where first element is resulting plaintext, and second element
-#  is the AES-GCM authentication tag
-#
-def decrypt_ciphertext(cipher_text, password, salt, iv, auth_tag)
-  encrypt = false
-  master_key = derive_key(password, salt)
-  aes_gcm(cipher_text, master_key, iv, encrypt, auth_tag)
-rescue OpenSSL::Cipher::CipherError, ArgumentError => e
-  terminate "Failed to derive cipher key. #{e.instance_of?(ArgumentError) ? e.message : 'Wrong password?'}"
-end
-
-#
-# Encrypt `plain_text` and return its ciphertext and authentication tag.
-#
-# @param [String] plain_text bytes to be encrypted
-# @param [String] password Backup file password as plaintext
-# @param [String] salt HMAC salt as bytes
-# @param [String] iv AES-GCM initialization vector as bytes
-#
-# @return [Array<String>] 2-element Array where first element is resulting ciphertext, and second element
-#  is the AES-GCM authentication tag
-#
-def encrypt_plaintext(plain_text, password, salt, iv)
-  encrypt = true
-  master_key = derive_key(password, salt)
-  aes_gcm(plain_text, master_key, iv, encrypt)
-rescue OpenSSL::Cipher::CipherError, ArgumentError => e
-  terminate "Failed to encrypt plaintext. #{e.instance_of?(ArgumentError) ? e.message : 'Invalid parameters?'}"
-end
-
-#
 # Prompt terminal user for password.
 # A drop-in replacement for $stdin.getpass for older Ruby versions.
 #
@@ -175,14 +96,13 @@ def getpass(prompt)
 end
 
 #
-# Decrypt vault with password from user input.
-# If successful, return plaintext vault data as JSON String.
+# Parse vault parameters from vault at `filename`.
 #
-# @param [String] filename Vault file to decrypt
+# @param [String] filename Vault filename
 #
-# @return [String] Plaintext vault as JSON String
+# @return [Hash] Vault parameters
 #
-def decrypt_vault(filename)
+def parse_vault_params(filename)
   begin
     obj = parse_json File.read(filename, :encoding => 'utf-8')
   rescue Errno::ENOENT => e
@@ -190,47 +110,10 @@ def decrypt_vault(filename)
   end
   cipher_text_with_auth_tag, salt, iv = extract_fields(obj).values_at(:cipher_text_with_auth_tag, :salt, :iv)
   cipher_text, auth_tag = split_cipher_text(cipher_text_with_auth_tag).values_at(:cipher_text, :auth_tag)
-
-  password = getpass('Enter 2FAS encrypted backup password: ')
-  plain_text, = decrypt_ciphertext(cipher_text, password, salt, iv, auth_tag)
-  parse_json(plain_text) # Ensure plain_text is valid JSON.
-  plain_text
+  { :cipher_text => cipher_text, :salt => salt, :iv => iv, :auth_tag => auth_tag }
 end
 
-#
-# Encrypt vault with given AES-GCM parameters.
-# If successful, return encrypted vault as JSON String.
-#
-# @param [String] plain_text Vault contents
-# @param [String] password Vault password
-# @param [String] salt HMAC salt as bytes
-# @param [String] iv Vault AES-GCM initialization vector as bytes
-# @param [String] reference_iv Reference AES-GCM initialization vector as bytes
-#
-# @return [String] Encrypted vault as JSON String
-#
-def encrypt_vault(plain_text, password, salt, iv, reference_iv)
-  cipher_text, auth_tag = encrypt_plaintext(plain_text, password, Base64.strict_decode64(salt),
-                                            Base64.strict_decode64(iv))
-  cipher_text_with_auth_tag = Base64.strict_encode64(cipher_text + auth_tag)
-  reference_cipher_text, reference_auth_tag = encrypt_plaintext(REFERENCE, password, Base64.strict_decode64(salt),
-                                                                Base64.strict_decode64(reference_iv))
-  reference_cipher_text_with_auth_tag = Base64.strict_encode64(reference_cipher_text + reference_auth_tag)
-  '{"services":[],"groups":[],"updatedAt":1708958781890,"schemaVersion":4,"appVersionCode":5000017,' \
-    '"appVersionName":"5.3.5","appOrigin":"android","servicesEncrypted":' \
-    "\"#{cipher_text_with_auth_tag}:#{salt}:#{iv}\"," \
-    '"reference":' \
-    "\"#{reference_cipher_text_with_auth_tag}:#{salt}:#{reference_iv}\"}"
-end
-
-#
-# Accept vault filename as a command-line argument, and optionally output format.
-# Decrypt the vault and write its contents to $stdout in specified output format.
-#
-# @param [String] filename Vault file to decrypt
-# @param [String] format Output format (Default: json)
-#
-def main
+def parse_args
   formats = %i[json csv pretty]
   options = { :format => :json, :except => [] }
 
@@ -261,8 +144,25 @@ def main
   rescue StandardError => e
     terminate "#{e}\n#{parser}"
   end
+  options
+end
 
-  plain_text = decrypt_vault(ARGV[0])
+#
+# Accept vault filename as a command-line argument, and optionally output format and fields to exclude.
+# Decrypt the vault and write its contents to $stdout in specified output format.
+#
+# @param [String] filename Vault file to decrypt
+# @param [String] format Output format (Default: json)
+#
+def main
+  options = parse_args
+  vault_params = parse_vault_params ARGV[0]
+  vault_params[:password] = getpass('Enter 2FAS encrypted backup password: ')
+
+  plain_text, = decrypt_ciphertext(vault_params[:cipher_text], vault_params[:password], vault_params[:salt],
+                                   vault_params[:iv], vault_params[:auth_tag])
+  parse_json(plain_text) # Ensure plain_text is valid JSON.
+
   $stdout.write case options[:format]
                 when :pretty
                   beautify remove_fields(entries_to_csv(plain_text), options[:except])
